feat: Driver for the QDevil QDAC
Driver for the QDevil QDAC. It is based on the “QDev\Qdac_channels.py” driver but is different in many ways as it leverages several of the latest developments from QDevil, and requires a QDAC from QDevil with at least firmware 1.07. The improvements include auto configuration, output calibration, and  synchronised multi-channel ramping.