feat: driver for Attocube ANC300
The test device has the ANC300 controller and two ANM150 axis
modules. Therefore no automatic test for other modules is
included in the driver (no tests possible).
The communication is done via PySerial. The ANC300sim.py is
included to simulate the correct answers from the device.
Activate the simulation with a flag in the ANC300.py file.

Update: Jupyter File set to no execution