Add appropriate lower limit for voltage compliance
Writing an SCPI command to set a voltage compliance value that is below
the minimum accepted by the device (0.2 mV) will cause the value to be
silently ignored, and the previous compliance level will stay. Without
this modification, it is possible to set the compliance to 0.1 mV but
accidentally pass 20 V, since that is the default voltage compliance
when the device is booted.